requirements.txt
pyproj
pyzmq
flask
pyaml
kafka-python